We start now to have finally a common base class for the Tx GUI
The method to display results is now common (except H317) all Txs
